pyproject.toml
[tool.poetry]
name = "course-airbyte-fraud-project"
version = "0.1.0"
description = "Marc Lamberti's project for the course on Airbyte"
authors = ["Thomaz Moreira <playjumanji>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
setuptools = "^69.2.0"
soda-core-snowflake = "^3.3.1"
dbt-snowflake = "^1.7.3"
python-nvd3 = "^0.15.0"
apache-airflow = "^2.9.0"
black = "^24.3.0"
apache-airflow-providers-airbyte = "^3.6.0"
astronomer-cosmos = "^1.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
